[Refactor] 파일명 리팩토링 - 컴포넌트가 아닌 파일은 .tsx에서 .ts 로 변경 - global.css 제거 (index.css 랑 중복) - store 이름 일관성 있게 수정 useAuth.ts -> authStore.ts
