+++ src/components/pages/SettingsPage.tsx
@@ -1,9 +1,9 @@
 import SettingsTemplate from '@/components/templates/SettingsTemplate';
-import useAuth from '@/store/useAuth';
+import useAuthStore from '@/store/authStore';
 import { useEffect } from 'react';
 
 const SettingsPage = () => {
-  const { isLoggedIn } = useAuth.getState();
+  const { isLoggedIn } = useAuthStore.getState();
 
   useEffect(() => {
     if (!isLoggedIn) window.location.href = '/login';
